move my diary detail after writing

+++ app/src/main/java/com/sopt/smeme/bridge/agent/mydiary/DiaryWriting.kt
@@ -6,7 +6,7 @@ interface DiaryWriting {
         targetCode: String,
         topic: Int,
         isPublic: Boolean,
-        onCompleted: () -> Unit = {},
+        onCompleted: (Long) -> Unit = {},
         onError: (Throwable) -> Unit = {}
     )
 }
